This step is the raw merge of the src folders.  **** WARNING!!! **** Trunk is not likely to build after this step.  DO NOT work with it at this point -- there are subsequent steps that will be performed to return Trunk to a stable (R2011a) code base!

git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@9513 222415a0-0dc2-4184-858d-a27cde375526

Former-commit-id: c590baafa1bbf2ffaec99f1b42374e7f9fa3f5af

+++ src/base/spacecraft/TextTrajectoryFile.cpp
@@ -1,8 +1,12 @@
-//$Header$
+//$Id$
 //------------------------------------------------------------------------------
 //                              TextTrajectoryFile
 //------------------------------------------------------------------------------
-// GMAT: Goddard Mission Analysis Tool
+// GMAT: General Mission Analysis Tool
+//
+// Copyright (c) 2002-2011 United States Government as represented by the
+// Administrator of The National Aeronautics and Space Administration.
+// All Other Rights Reserved.
 //
 // ** Legal **
 //
@@ -93,9 +97,9 @@ bool TextTrajectoryFile::ReadLine()
         mInStream >> mTempData[i];
 
     mTrajData.time = mTempData[0];
-    mTrajData.x = mTempData[1];
-    mTrajData.y = mTempData[2];
-    mTrajData.z = mTempData[3];
+    mTrajData.x = (float)mTempData[1];
+    mTrajData.y = (float)mTempData[2];
+    mTrajData.z = (float)mTempData[3];
     
     // add to mData
     mDataArray.push_back(mTrajData);
